test: split test_index.cpp into focused files
Keep the existing test coverage intact by regrouping TestIndex methods into focused translation units and sharing common helpers through a private header.
